+++ src/db/helpers/index.ts
@@ -1,0 +1,6 @@
+import { createDatabaseError } from "./errors";
+
+export {
+  // eslint-disable-next-line import/prefer-default-export
+  createDatabaseError,
+};
