+++ src/index.ts
@@ -71,6 +71,12 @@ const envVarsOk = verifyEnvVars([
   {
     envVar: "JWT_PUBLIC_KEY",
   },
+  {
+    envVar: "ACCOUNT_VERIFICATION_TIME",
+    requirements: {
+      type: "number",
+    },
+  },
 ]);
 
 if (!envVarsOk) {
