+++ src/server/routes/v1/routes/admin/index.ts
@@ -1,0 +1,7 @@
+import { FastifyInstance } from "fastify";
+
+import users from "./users";
+
+export default async (fastify: FastifyInstance): Promise<void> => {
+  fastify.register(users, { prefix: "users" });
+};
