+++ src/server/routes/v1/routes/index.ts
@@ -1,0 +1,9 @@
+import admin from "./admin";
+import user from "./user";
+import visitor from "./visitor";
+
+export {
+  admin,
+  user,
+  visitor,
+};
