+++ src/server/routes/v1/routes/visitor/bandsy/recover.ts
@@ -1,0 +1,11 @@
+import { FastifyInstance } from "fastify";
+
+export default async (fastify: FastifyInstance): Promise<void> => {
+  fastify.get("/", async () => {
+    const rawrxd = "rawrxd";
+
+    return {
+      rawrxd,
+    };
+  });
+};
