+++ src/server/routes/v1/routes/visitor/index.ts
@@ -1,0 +1,9 @@
+import { FastifyInstance } from "fastify";
+
+import bandsy from "./bandsy";
+import oauth from "./oauth";
+
+export default async (fastify: FastifyInstance): Promise<void> => {
+  fastify.register(bandsy, { prefix: "bandsy" });
+  fastify.register(oauth, { prefix: "oauth" });
+};
