+++ src/server/routes/v1/routes/visitor/oauth/authenticate.ts
@@ -1,0 +1,11 @@
+import { FastifyInstance } from "fastify";
+
+export default async (fastify: FastifyInstance): Promise<void> => {
+  fastify.get("/", async () => {
+    const rawrxd = "rawrxd";
+
+    return {
+      rawrxd,
+    };
+  });
+};
