+++ src/server/routes/v1/routes/visitor/oauth/index.ts
@@ -1,0 +1,7 @@
+import { FastifyInstance } from "fastify";
+
+import authenticate from "./authenticate";
+
+export default async (fastify: FastifyInstance): Promise<void> => {
+  fastify.register(authenticate, { prefix: "authenticate" });
+};
