+++ src/server/routes/v1/types/codes.ts
@@ -1,0 +1,13 @@
+enum BandsyResponseCodes {
+  ERROR_HANDLER_ERROR,
+  UNKNOWN_ERROR,
+  VALIDATION_ERROR,
+  CLIENT_ERROR,
+  SERVER_ERROR,
+  DUPLICATE_EMAIL,
+}
+
+export {
+  // eslint-disable-next-line import/prefer-default-export
+  BandsyResponseCodes,
+};
