+++ src/server/routes/v1/types/errors.ts
@@ -1,0 +1,11 @@
+import { FastifyError } from "fastify";
+import { BandsyResponseCodes } from "./codes";
+
+interface IBandsyError extends FastifyError {
+  bandsyCode: BandsyResponseCodes;
+}
+
+export {
+  // eslint-disable-next-line import/prefer-default-export
+  IBandsyError,
+};
