+++ src/server/routes/v1/types/http.ts
@@ -1,0 +1,11 @@
+enum HttpResponseCodes {
+  OK = 200,
+  OK_NO_CONTENT = 204,
+  CLIENT_ERROR = 400,
+  SERVER_ERROR = 500,
+}
+
+export {
+  // eslint-disable-next-line import/prefer-default-export
+  HttpResponseCodes,
+};
