+++ src/server/routes/v1/types/index.ts
@@ -1,0 +1,9 @@
+import { HttpResponseCodes } from "./http";
+import { BandsyResponseCodes } from "./codes";
+import { IBandsyError } from "./errors";
+
+export {
+  HttpResponseCodes,
+  BandsyResponseCodes,
+  IBandsyError,
+};
